requirements.txt
pyTelegramBotAPI==3.7.3
python-dotenv==0.14.0
requests==2.24.0
httplib2==0.18.1